+++ tomcat/src/main/java/org/apache/coyote/http11/session/SessionManager.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.auth;
+package org.apache.coyote.http11.session;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
@@ -7,9 +7,6 @@ import org.apache.catalina.Manager;
 public class SessionManager implements Manager {
 
     private static final Map<String, Session> SESSIONS = new ConcurrentHashMap<>();
-
-    public SessionManager() {
-    }
 
     @Override
     public void add(final Session session) {
